+++ NEWS.md
@@ -1,3 +1,9 @@
+# arrR 0.2.3
+
+* Resulting object contains parameters and starting values
+* `get_limits` can return limits over time period
+* `plot.mdl_rn` also takes limits if `summarize = TRUE` for seafloor
+
 # arrR 0.2.2
 
 * Better plotting labels